requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-dotenv==1.0.0
openpyxl==3.1.2
requests==2.31.0
httpx==0.27.0
